adding regression to client
