+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "tabpfn-client"
-version = "0.0.13"
+version = "0.0.14"
 requires-python = ">=3.10"
 dependencies = [
     "httpx>=0.24.1",
