+++ tabpfn_client/__init__.py
@@ -1,4 +1,5 @@
-from tabpfn_client.tabpfn_classifier import init, TabPFNClassifier
+from tabpfn_client.config import init, reset
+from tabpfn_client.estimator import TabPFNClassifier, TabPFNRegressor
 from tabpfn_client.service_wrapper import UserDataClient
 
-__all__ = ["init", "TabPFNClassifier", "UserDataClient"]
+__all__ = ["init", "reset", "TabPFNClassifier", "TabPFNRegressor", "UserDataClient"]
