+++ tabpfn_client/prompt_agent.py
@@ -4,7 +4,7 @@ from password_strength import PasswordPolicy
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from tabpfn_client.tabpfn_classifier import UserAuthenticationClient
+    from tabpfn_client.estimator import UserAuthenticationClient
 
 
 class PromptAgent:
