+++ tabpfn_client/service_wrapper.py
@@ -1,5 +1,6 @@
 import logging
 from pathlib import Path
+from typing import Literal
 
 from tabpfn_client.client import ServiceClient
 from tabpfn_client.constants import CACHE_DIR
@@ -179,7 +180,7 @@ class InferenceClient(ServiceClientWrapper):
 
         self.last_train_set_uid = self.service_client.upload_train_set(X, y)
 
-    def predict(self, X, config=None):
+    def predict(self, X, task:Literal["classification","regression"], config=None):
         return self.service_client.predict(
-            train_set_uid=self.last_train_set_uid, x_test=X, tabpfn_config=config
+            train_set_uid=self.last_train_set_uid, x_test=X, tabpfn_config=config, task=task,
         )
